[general] Refactored backends, renderers, added converters #70
This is the first step in implementing #70.

This patch does not introduce changes of behaviour in the library,
infact tests remain unaltered, except for their names (because they
were named after renderer classes which now are a much smaller
part of the library).

This patch introduces the concept of converter classes, which are
responsible for converting the NetJSON configuration dictionary
to an intermediate data structure that is then used by renderers
to be rendered.

The backward conversion will then implement the reverse process:

Parsers will be responsible of converting native configuration
to the intermediate data structure;
Converters will have a to_netjson method that will be responsible
of converting the intermediate data structure to the NetJSON
configuration dictionary.