added external backend querying at runtime
this will use the entry_points feature of pkg_resources to query other
installed packages for additional backends.

- [doc] added page create_your_backend
- prevent import errors from crashing the cli

during load the module will be imported but if an ImportError
is raised the cli crash. This will prevent the cli from crashing
but will not expose the backend to the user. An error message
will be logger to console

- exclude except form from flake qa

this form is the only one compatible between python2 and python3
but flake is also worried about the e variable which is not used