[openwrt] Fixed TypeError caused by type mismatch
Fixed a TypeError that was raised due to a concatenation (`ipv4 + ipv6`) between two addresses that were of different types, `list` and `str`. I resolved this by applying the same type conversion that was applied to the `ipv4` variable: If only one address is found, convert it to a list of a single element.